feat: extract help from feedback
